+++ apps/web/src/app/nonce/layout.tsx
@@ -1,0 +1,8 @@
+import { ThemeProvider } from "next-server-theme/dev/server.ts";
+import { headers } from "next/headers";
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+    const nonce = headers().get("x-nonce");
+
+    return <ThemeProvider nonce={nonce}>{children}</ThemeProvider>;
+}