+++ apps/web/src/middleware.ts
@@ -30,3 +30,7 @@ export function middleware(request: NextRequest) {
 
     return response;
 }
+
+export const config = {
+    matcher: "/**/nonce"
+};